Implement book editor layout and enhance block editing features
- Added new CSS styles for the book editor layout, including toolbar and workspace components.
- Updated JavaScript for block editor to normalize Unicode text and improve textarea handling.
- Enhanced bounding box overlay functionality to scale boxes based on image dimensions.
- Introduced new utility functions for parsing and normalizing bounding box data.
- Improved the replica view to support dynamic scaling of bounding boxes and text content.
- Updated templates to integrate new book editing features and improve user experience.

+++ kalanjiyam/utils/ocr_types.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass
 @dataclass
 class OcrResponse:
     text_content: str
-    bounding_boxes: list[tuple[int, int, int, int, str]]
+    bounding_boxes: list[tuple[float, float, float, float, str]]
     layout_html: str | None = None
     blocks: list[dict] | None = None
     content_format: str = "plain"
@@ -78,7 +78,7 @@ def post_process(text: str) -> str:
 
 
 def serialize_bounding_boxes(
-    engine: str, boxes: list[tuple[int, int, int, int, str]]
+    engine: str, boxes: list[tuple[float, float, float, float, str]]
 ) -> str:
     if not boxes:
         return ""
